+++ src/main/java/com/example/momo/global/infrastructure/springEvent/payment/PaymentRefundedEvent.java
@@ -1,4 +1,4 @@
-package com.example.momo.global.infrastructure.springEvent;
+package com.example.momo.global.infrastructure.springEvent.payment;
 
 import java.time.LocalDateTime;
